+++ src/vision/processor.py
@@ -4,7 +4,7 @@ from collections.abc import Iterable
 
 
 class VisionProcessor:
-    def process(self, stream: Iterable[tuple[MatLike, float]]) -> Iterable[tuple[MatLike, float]]:
+    def process(self, stream: Iterable[tuple[MatLike, int]]) -> Iterable[tuple[MatLike, int]]:
         for frame, timestamp in stream:
             flipped_frame = cv2.flip(frame, 1)
             yield flipped_frame, timestamp
